+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * Properly escape special chars in output XML. Allow control chars in properties.
   * Renamed `UpdateHandler::connect` to `UpdateHandler::openDownloadStream`
   * Added 2 more `Configuration.Builder::signer` overloads.
+  * Default bootstrap now deletes old filesif `--syncLocal` was used.
 * **1.3.2** *— h/t [@ChristianCiach](https://github.com/ChristianCiach)*
   * Control how files are downloaded with `UpdateHandler::connect`.
   * Delete old files with `Configuration::deleteOldFiles`.